+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Implemented Client#calendar.
+- Implemented Client#clock.
 
 ## [0.1.0] - 2019-07-04
 ### Added